Add rate limiter

+++ app/config/bot_config.go
@@ -4,5 +4,5 @@ type BotConfig struct {
 	TwitchOAuthToken string `json:"twitch_oauth_token"`
 	BotUsername      string `json:"bot_username"`
 	TargetChannel    string `json:"target_channel"`
-	RateLimit        uint8  `json:"rate_limit"`
+	RateLimit        int    `json:"rate_limit"`
 }
